Housekeeping - requestlib test tidy up
Having clocked that we hadn't set the `{ retry: shortBackoffLimitRetryOptions }` in our new retry tests we added them and saw run times improve. But then we realised this made needing to override the Lab test timeout redundant so we remove that here.

Also, whilst using the timeout scenarios as the template for updating the network issue ones, we realised that the timeout tests were not checking that a log of the error is record after all retries fail.

So, we've gone back and added them in.