Stop showing supp source charge as negative
https://eaflood.atlassian.net/browse/WATER-4363

We've just dealt with [Fix bill licence supp source decimal in wrong place](https://github.com/DEFRA/water-abstraction-system/pull/714). This caused us to take a closer look at how the supported source charge is displayed in the new views.

Our eagle eyed testers spotted we were showing the value as a negative if the transaction itself is a credit and whether this is correct. When we double checked the legacy view we found it's not! 🤦

This change fixes the display so the supported source charge is always displayed as a positive.