Make requests to the CHA more resilient (#718)
https://eaflood.atlassian.net/browse/WATER-4365

As part of looking at re-implementing the SROC annual billing engine in this project, we found a solution that allows us to process bills in parallel (will be added in later commits!) However, the more we do the more strain this puts on the [Charging Module API](https://github.com/DEFRA/sroc-charging-module-api) to the point we start seeing timeouts and `ECONNRESET` errors.

With just a few tweaks on our side, we could increase our tolerance of these errors. There is still a point where we are pushing it too hard. But the more throughput we get, the faster we can complete the bill run process.

So, this change does the following.

- **Add charging module specific request timeout** We don't just want to increase the timeout for all web requests. Some impact the experience of the end-user and they shouldn't be made to watch a page spinning for longer than necessary. So, we had a specific one for the Charging Module
- **Retry `ECONNRESET` errors.** We learnt this can happen when an API is under pressure. It doesn't mean the Charging Module can't handle the next attempt so it is safe to try these again.

---

Also a bit of housekeeping. Whilst using the timeout scenarios as the template for updating the network issue ones, we realised that the timeout tests were not checking that a log of the error is recorded after all retries fail.

So, we've gone back and added them in.