Add charge data to licence review page
https://eaflood.atlassian.net/browse/WATER-4219

This PR is to amend an existing page the review licence page. This is for a two part tariff bill run. The first half of this page was completed in a separate PR due to the ticket being split. This PR is to add the charge data for the page and the licence status.