Add a new wipe.database.js module
In order to invoke this from `npm run` and our `package.json` we need a module that calls rather than exports the function we want to run. In our case it is an async function so we also need to use some Node wizardry to await a top level async function.

This new module is what enables both these things to happen.