Add wipe test DB function to project (#724)
https://eaflood.atlassian.net/browse/WATER-4365

> For context this came out of us working on re-implementing the SROC annual bill run using what we've learnt and components from our supplementary billing engine.

Whilst spiking on getting a working SROC annual billing engine we found that some of the legacy tables the migrations were creating in the test DB didn't match with the real thing. No doubt a result of our original intent being legacy migrations for the test DB did the minimum needed to get tests passing.

We could add fix scripts, but as this is just for our test-db we want to avoid starting to make our migrations folder even more noisy with fix-migrations.

So, we'd rather go back and fix them. The only problem is this then breaks running migrations for the test DB. When we've done this before we can resolve the issue by manually going in and deleting everything first.

💡💡💡

That's when @robertparkinson had the great idea of why don't we just delete everything anyway!? If we do this, whether we've broken them or not the test DB should always successfully be built and work.

So, this change adds a new function that will delete the schemas, views and tables from the `wabs_system_test`. We'll then update the `migrate:test` npm run script to call this first. 🤞 after this test migrations will never fail again!