Add no returns to licence review page (#731)
https://eaflood.atlassian.net/browse/WATER-4191

Following on from testing for the above ticket it was discovered that it is indeed possible to have a licence in the two-part tariff bill run with no returns. Initially we didn't think this was a possibility but bad data does slip through often enough to warrant us to add in logic to handle it. This PR is for the licence review page to handle no returns, adding in additional text and allowing the licence to still be clicked into and not error.