Remove abstraction text for under query licence (#736)
https://eaflood.atlassian.net/browse/WATER-4191

During the testing for the review licence page in a twp part tariff bill run, it was discussed in a blocker review the text for a return that is under query. Currently if a return is under query it shows a text box stating the return is either fully abstracted or over abstracted. This is wrong as a return that is under query is not allocated on the charge element and therefore does not meet any of those definitions. This change is to remove that text box for returns under query. This change also includes content changes on the page for unmatched returns. This is since an unmatched return would not be allocated to a charge element so we are removing the allocated text field.