Fix broken tests after water.billing_trans update
We've add a number of constraints that exist on the real table but which we didn't have in our test DB. The majority are focused on ensuring certain fields are not null if the type is 'x' or the scheme is 'y' for 'reasons'!

The key one is that they should all have the start and end date populated which would be based on the charge period. In most cases this aligns with the billing period so we've created a helper that returns the current financial year's start and end dates rather than hard code in values on the basis the assumption for most tests are that we are working in the current financial year.