Start tidy up - update filenames
For all the migrations the timestamps ensure that as migrations are added their order is retained. This is important as DB changes are made overtime to a system.

But our legacy migrations are purely for our test DB. Which means we can alter the timestamps to allow us to control the order of when things happen. Couple this with the fact we now always wipe before running migrations against the test DB means we can also start to group tables by schema. This should make it easier to see what has and hasn't been replicated.

SO, this is the first change in a series of tidy ups for the legacy migrations folder.