Remove Landing Page From System
As part of a Junior Dev Project, Roble was tasked with creating an online form for the current water abstraction licence paper form.

This was previously expected to be used as part of a future implementation for external users to sign up for a licence, but this has now been declared outside the scope.

This PR is focused on removing any of the changes from this previous ticket (https://eaflood.atlassian.net/browse/WATER-4359) and PR (Create Landing page #728).