Refactor where billLicenceId is applied
Previously, we passed the billLicence to `SendTransactionsService`, extracted the ID and applied it to the transaction along with the values returned from the Charging Module.

We don't rightly know why we chose to do this there (most likely the result of a hack we never refactored) but it feels out of place.

Better to have it applied to the transaction at the point we generate the transaction.