Move SendTransactions to root with refactoring (#745)
https://eaflood.atlassian.net/browse/WATER-4365

> For context this came out of us working on re-implementing the SROC annual bill run using what we've learnt and components from our supplementary billing engine.

As part of looking at re-implementing the SROC annual billing engine in this project our spike (WATER-4348 ) confirmed we'd need to reuse some of the services currently sitting in `app/services/bill-runs/supplementary`.

We moved most of these in [Move shared billing services to bill-runs root](https://github.com/DEFRA/water-abstraction-system/pull/720) but left `app/services/bill-runs/supplementary/send-transactions.service.js` out even though we also need to reuse it.

When working on the spike we found there is some tidying up we can do in the service. We also spotted we were passing in `BillLicence` purely to set the `billLicenceId`. We now feel this should be done elsewhere.

So, we're doing those changes here as they are a little more involved than simply moving the service.