Add annual bill run FetchBillingAccountsService
https://eaflood.atlassian.net/browse/WATER-4365

> For context this came out of us working on re-implementing the SROC annual bill run using what we've learnt and components from our supplementary billing engine.

As part of looking at re-implementing the SROC annual billing engine in this project our spike (WATER-4348 ) confirmed we could simplify the implementation and improve performance if we extracted the billing account details at the same time as the charge versions to be billed.

In supplementary billing we first use `app/services/bill-runs/supplementary/fetch-charge-versions.service.js` to fetch all the charge versions that should be considered for billing. Then we call `app/services/bill-runs/supplementary/fetch-billing-accounts.service.js` to get the billing account info and merge the two using `PreGenerateBillingDataService` to pre-generate the bare bones bills and bill licences we need.

In `app/services/bill-runs/supplementary/process-billing-period.service.js` we then have some 'clever' code that builds the billing data we need from the charge versions for each bill and bill licence we pre-generated.

It took us awhile to get back up to speed on all this which made us review how we were doing things and come to the conclusion; what if we started at the bill level?

What if rather than fetching the charge versions to consider and then extracting from this the bills we need to generate we instead go straight to fetching the billing accounts to consider? With our new [relationship between BillingAccountModel & ChargeVersionModel](https://github.com/DEFRA/water-abstraction-system/pull/750) we can directly associate the charge versions we need to the billing accounts we fetch.

This strips away the need for any pre-generation of data or calls to methods like `_buildBillingDataWithTransactions()` and `buildDataToPersist()`.

So, this change adds the new service ready for use in our coming annual billing engine.