Add Annual ProcessBillingPeriodService
https://eaflood.atlassian.net/browse/WATER-4365

> For context this came out of us working on re-implementing the SROC annual bill run using what we've learnt and components from our supplementary billing engine.

The main component of our supplementary billing engine is its `ProcessBillingPeriodService`. But it has to deal with generating bills across a 6 year period as well as checking what has previously been billed and comparing that with what is due to be billed.

Fortunately, the annual billing engine doesn't have any of these additional complications. Which means we can't reuse what we built for supplementary billing.

So, this adds a `ProcessBillingPeriodService` for annual billing which will be responsible for fetching the billing accounts to bill, and then creating their bills by generating the transactions and sending them to the [Charging Module API](https://github.com/DEFRA/sroc-charging-module-api).

Once in place we can hook up the create bill request to this engine and start auditing if what we've built is correct.