Add new billing config
Specifically this allows us to configure in each environment how many billing accounts are processed at the same time. For now there is only config for annual billing. But we hope to bring the lessons we've learnt building the annual billing engine to our existing supplementary engine, and the future 2PT annual and supplementary engines. Which means they'll need similar config that has a home here.