Handle live bill run check for annual bill runs (#761)
https://eaflood.atlassian.net/browse/WATER-4379

When we built the supplementary billing engine one of the things we were required to do was stop a user from creating a second bill run for the same region if one was already 'live'.

But supplementary differs in that you can create as many bill runs as you like within the same billing period (financial year). Annual bill runs only 1 per billing period is allowed.

Our existing `CheckLiveBillRunService` is handling when an annual bill run with a status of 'processing', 'ready', 'review', or 'queued' exists.

But if a 'sent' annual bill run exists it should be rejecting the request (bill run is live!) Instead, it is telling the app it can go ahead and create a new one.

This change updates the service to handle checks for annual billing as well as supplementary.