Update annual process bill run to connect engine (#760)
https://eaflood.atlassian.net/browse/WATER-4379

> For context this came out of us working on re-implementing the SROC annual bill run using what we've learnt and components from our supplementary billing engine.

We've been making a series of changes to our services in `app/services/bill-runs` in preparation for supporting creating annual bill runs. Primarily this has been refactoring services built for supplementary to be reused for annual billing. But it also involved adding the heart of the annual billing engine; `app/services/bill-runs/annual/process-billing-period.service.js`.

Now we are ready to connect these changes to the request coming. That connection is managed by `app/services/bill-runs/annual/process-bill-run.service.js` which currently is just a placeholder. This change updates it to handle orchestrating the billing process for annual bill runs.