Fix annual billing not excluding workflow licences
https://eaflood.atlassian.net/browse/WATER-4379

We are working to replace the legacy SROC annual billing engine with one in this project that exploits what we did with SROC supplementary billing.

First pass of testing has highlighted a discrepancy with the bill amounts. When we looked into it we found the new engine was including licences that have records in workflow.

Any licence with un-deleted workflow record should be excluded from billing. This change fixes the issue.