First pass of fixing broken unit tests
When we reviewed the unit tests to see how this got through (because we are testing for it!) we spotted that the test charge versions we are creating were left to generate their billing account ID. But the query filters by matching the billing account ID for the account we've fetched against the charge versions on the linked licences.

So, first fix is to ensure our test charge versions also feature the same billing account ID as our test account. This exposed we'd got 2 other tests wrong so we also fix those.