Hmm. Test seems right, Breaking on purpose
With the billing account ID now being set the test still passed. We were expecting it to fail as this would reflect the results we see in the annual bill run.

Putting in a different ID causes the test to fail so we're still not sure exactly what is going wrong. :-(