Ah ha! Put test back as is correct
We have figured out the problem!

If a billing account is linked to one licence and that licence is in workflow then we should ignore it from our results. Hence the test with the billing account ID is correct.

The scenario that led to a discrepancy in the bill run is different though.

It was where a billing account is linked to 2 licences, only one of which is in workflow. That means in our engine we include the billing account (which is correct).

What we have failed to do is then discount the licence in workflow from the data we are using to generate that accounts bill.