Persist match and allocate issues and status
https://eaflood.atlassian.net/browse/WATER-4378

During the refinement of future two-part tariff tickets it was noted that the status of a licence could be changed by the user from a 'review' status to a 'ready' status. Before this was noted we were working out the licence status based on the issues the licence has with its returns and charge elements. This one done on every page where the licence status is being shown. Now that we know a status can be overwritten regardless of its issues, we now need to persist the licence status during the first time the engine works it out and show the persisted status instead of working it out each time. This will allow users to manually change the status.
The same thing applies for the charge element. The status of the element is determined by the issues on it. This status at the moment is worked out every time it is displayed on a page. We have since learnt through refinement that a user can allocate volumes on the element and this will change the status from 'review' to 'ready'. This is regardless of the issues that still remain on a charge element. So with this new information it has made it clear that we now need to persist the initial status of the charge element to allow this to be overwritten when needed.
This opened up a lot of discussions about how we could refactor the review tables to work more effectively with how the two-part tariff engine works now. When we first built the tables we did so with the knowledge we had at the time. Now that we have worked on more tickets and started building actual pages we realised we could persisted the data better.
This PR covers the above along with some refactoring to our engine.