Fix annual billing not excluding workflow licences (#763)
https://eaflood.atlassian.net/browse/WATER-4379

We are working to replace the legacy SROC annual billing engine with one in this project that exploits what we did with SROC supplementary billing.

The first pass of testing has highlighted a discrepancy with the bill amounts. When we looked into it we found the new engine was still including licences that have records in workflow as part of the calculation.

Our new engine is only returning billing accounts that have applicable charge versions. What we failed to do was then filter again the charge versions returned for each billing account.

The scenario we found was a billing account linked to 2 licences, one of which was in workflow. The billing account was included because it has a charge version linked to a licence that isn't in workflow.

But then we fetched for the billing account _all_ the charge versions including the one linked to a licence in workflow.

We should be doing the same filtering in both places. This fix ensures that.