Fix bill licences with no transactions in bill run (#765)
https://eaflood.atlassian.net/browse/WATER-4379

We are working to replace the legacy SROC annual billing engine with one in this project that exploits what we did with SROC supplementary billing.

The first pass of testing has highlighted a discrepancy with how many licences are being included in some bills. For example, the legacy bill run will only display 1 but the new engine will display 3.

When we looked into it we found it was because only 1 of the 3 licences had applicable transactions. Two of the licences had no billable days.

The legacy is correct in this case, we shouldn't be creating bill licence records where there were no transactions. This change fixes the annual billing engine to deal with this scenario.

> For reference the licences had no billable days because they were revoked before the abstraction period on the charge reference i.e. revoked before the abstraction period so the charge period is null.