Update allocation engine to handle overdue returns (#759)
https://eaflood.atlassian.net/browse/WATER-4377

Update the match & allocation engine to ‘allocate’ overdue returns.

The complication is the engine is built on matching return submission lines to charge elements by comparing the line’s start and end dates to a charge element’s calculated abstraction dates (amongst a host of other parameters).

So where there is a licence with an overdue return that has been matched to an element, the engine needs to ensure that the customer is charged the authorised amount for that element.

The engine is to be updated so that after all completed returns have had their volumes allocated. If there are any overdue returns, the engine will attempt to match those to a charge element. If matched the engine will then bump the volume allocated to that element up to its authorised volume.