Display public water supply charge in bill licence
https://eaflood.atlassian.net/browse/WATER-4392

When you view a bill licence we display the transactions for that licence in the current bill. If there are any additional charges there is a section that appears that will list these out.

Billing & Data have highlighted that when the additional charge is 'Supported source' we handily show the charge in brackets afterwards. If the charge is 'Public water supply' we don't.

This change makes it so!