Add Bill Run Charge Version Years model (#777)
https://eaflood.atlassian.net/browse/WATER-4387

We need to add this to support our work to improve how bill runs are cancelled. Basically, replace the crummy legacy one!

The driving reason is to make it visible to users when a bill run is being cancelled as this impacts performance. If we can make cancelling bill runs visible, users can make more informed decisions about when to start creating a new one.

In support of that, we need to add the `BillRunChargeVersionYearModel` as it's the one table we don't have a model for and which we need to delete stuff from when cancelling a bill run.