Alter `return_logs` view https://eaflood.atlassian.net/browse/WATER-4378
Whilst working on two-part tariff we have noticed that the column in the `return_log` table called `return_requirement` is called "Return reference" in the service.

We are therefore going to rename the `return_requirement` column in the view to `return_reference` so that it is the same as in the service.