Add submit cancel bill run endpoint (#782)
https://eaflood.atlassian.net/browse/WATER-4387

We need to add this to support our work to improve how bill runs are cancelled. Basically, replace the crummy legacy one!

In [Add cancel bill run page](https://github.com/DEFRA/water-abstraction-system/pull/780) we add the page where the user confirms they want to cancel (delete) the bill run.

This final step implements the service that will do the cancelling and the endpoint to trigger it.

Unlike the legacy version when we cancel a bill run, we'll

- immediately update the bill run status to `cancel`
- redirect the user to the bill runs page
- continue the process of cancelling the bill run in the background

This way the user should never see a timeout waiting for a bill run to cancel ever again. It also puts the service into a state where we can start to display to users that a bill run is in the process of being cancelled, the end goal of WATER-4387.