Migrate view empty bill run page
https://eaflood.atlassian.net/browse/WATER-4396

This change migrates the handling of viewing an empty bill run from the legacy [water-abstraction-ui](https://github.com/DEFRA/water-abstraction-ui) to this project.

Because we've migrated the cancelling of a bill run to this project (WATER-4387) we need to update the legacy pages which have a cancel button to point to it.

Or, as we're choosing to do, we straight up replace them with our own!

This change deals with when the bill run has a status of empty and the page that gets shown.