Refactor duplicated functions to base presenter
If we pressed straight ahead with an empty bill run presenter we'd be repeating a number of the functions for the 3rd time.

So, abiding by the rule of 3 it is time to refactor them into something that can be reused. So, we move the functions into our base presenter ready for us to update the presenters that use them.