Migrate view empty bill run page (#783)
https://eaflood.atlassian.net/browse/WATER-4396

This change migrates the handling of viewing an empty bill run from the legacy [water-abstraction-ui](https://github.com/DEFRA/water-abstraction-ui) to this project.

Because we've migrated the cancelling of a bill run to this project (WATER-4387) we need to update the legacy pages with a cancel button to point to it.

Or, as we choose to do, we replace them with our own!

This change deals with when the bill run has a status of empty and the page that gets shown.

---

This change was the tipping point of several bill-run-related presenters duplicating the same functions 3 times. As per the [rule of three](https://en.wikipedia.org/wiki/Rule_of_three_(computer_programming)) we refactor them to the `base.presenter.js` to make them reusable.