Fix badge for empty status
https://eaflood.atlassian.net/browse/WATER-4396

Unfortunately, we only spotted after [migrating the view empty bill run page](https://github.com/DEFRA/water-abstraction-system/pull/783) from the legacy service to this one that the status badge was wrong.

It should be styled as a grey badge (inactive) but we were displaying it as blue (info). Fixing it required adding an if/else check into 2 views which made us re-think how we were doing the bill run status badges.

So, in this change we fix the badge for `empty` and along the way we add a new `statusBadge()` macro that moves the logic of converting bill run status into badge type to one place.