Migrate view errored bill run page
https://eaflood.atlassian.net/browse/WATER-4396

This change migrates the handling of viewing an errored bill run from the legacy [water-abstraction-ui](https://github.com/DEFRA/water-abstraction-ui) to this project.

Because we've migrated the cancelling of a bill run to this project (WATER-4387) we need to update the legacy pages with a cancel button to point to it.

Or, as we choose to do, we replace them with our own!

This change deals with when the bill run has a status of `error` and the page that gets shown.