Remove unused frontend macros from cancel view
https://eaflood.atlassian.net/browse/WATER-4387

We spotted a little to late that the `app/views/bill-runs/cancel.njk` view is importing macros it is not using. We know this was because of a copy and paste and doesn't break anything.

Still, we don't like to make a habit of importing things and then not using them.