Remove unused frontend macros from cancel view (#786) (#787)
https://eaflood.atlassian.net/browse/WATER-4387

To make cancelling bill runs more visible in the service we needed to re-implement the cancel bill run feature. Because we've also replaced all the legacy bill run pages we didn't think we'd need to deal with any legacy views.

That was until we spotted that the PRESROC two-part tariff review page is still alive and kicking in the service! You can cancel a bill run from that page so we must support going back to it in our confirm cancel bill run page.

This change makes that so.