Cancel a 2PT Bill Run from the Review Licences Screen
https://eaflood.atlassian.net/browse/WATER-4189

Add a "Cancel bill run" button to the "Review licences" screen which was built in a previous PR https://github.com/DEFRA/water-abstraction-system/pull/663

On clicking the cancel button the user is taken to a confirmation screen displaying the header details of the bill run and a confirmation button to cancel the bill run.

Upon clicking the confirmation button the bill run should be deleted from the Charging Module & database, along with the data that has been persisted following the matching process.