Trigger two-part tariff annual bill run spike
https://eaflood.atlassian.net/browse/WATER-4375

In the legacy service there is no clear separation between two-part tariff annual billing and two-part tariff supplementary. The previous team, for the right reasons, tried to cover it all in one. Hindsight has shown it would be better to handle them separately. This is a wish from users as well as those of us who have to build it!

This means we are going to need to _update_ the 'Create a bill run' journey to handle the changes. And if you haven't guessed from how we do things here, what this _really_ means is we get to replace the whole thing! 😁

So, we intend to use this branch to spike out a working version. We'll then break up the changes into smaller PR's for review once we know how we want to do things.