WIP completed journey
This is a 'working' journey for creating a new bill run.

Validation is working for each of the pages and the user's choices are being saved to the session and replayed if they go forwards and backwards.

What it's not doing is branching based on choices. For example, if you select annual or supplementary you only then need to select a region.

We also need to determine the financial years to display for two-part tariff.

Finally we'll need to trigger the bill run(s) at the end of the journey.