Simplify journey for two-part tariff
We're no longer going to try and be clever with determining the year. Displaying the year is only temporary any way until they have caught up with the backlog and got all the outstanding 2PT bills out.

After that, like supplementary and annual there would never be a need to pick the financial year or the season.