Direct end journeys through new /generate endpoint
So, we now have a /generate endpoint where the other handlers will direct their requests when a users gets to the end of a create bill run journey.

Mirroring what the existing service does we then check to see if the requested bill run can be created. If we find a matching bill run exists we redirect the user to a page that displays its details.

If we can generate a bill run ... well that's the next step!