Move non-model helpers out of helpers
We created `test/support/helpers` initially a place to put small helper functions. When models came in we started creating helper modules to quickly create records in the DB for each model. These were kept separate from the general helpers because we put them in folders named after their schemas.

But when we switched to building our models based on views schemas became meaningless and all the model helpers were moved into the same folder as the general helpers. They became lost.

We had a discussion amongst the team and consensus was to move the 'non-model' helpers to the root of `test/support` and drop 'helper' from their name. Now, whenever you see `*.helper.js` we know it's a model helper and our more generic modules are no longer lost.