Move non-model helpers out of helpers (#795)
We initially created `test/support/helpers` as a place to put small helper functions. We started creating helper modules to quickly create records in the DB for each model when models came in. These were kept separate from the general helpers because we put them in folders named after their schemas.

But when we switched to building our models based on views schemas became meaningless and all the model helpers were moved into the same folder as the general helpers. They became lost.

We discussed this as a team and the consensus was to move the 'non-model' helpers to the root of `test/support` and drop 'helper' from their name. Now, whenever you see `*.helper.js` we know it's a model helper and our more generic modules are no longer lost.