Add defra-user-id header to legacy requests (#796)
https://eaflood.atlassian.net/browse/WATER-4375

> Part of a series of changes related to replacing the create bill run journey to incorporate changes for two-part tariff

If we are to take over the create bill run journey it means _we_ are now responsible for triggering legacy bill runs rather than vice versa. That means we need to trigger the [water-abstraction-service](https://github.com/DEFRA/water-abstraction-service) `POST /water/1.0/billing/batches` endpoint.

Simple, we thought! We already have `LegacyRequestLib` that adds support for talking to the other services from our project. That was until we kept getting `403` errors. After some digging, we found the cause.

Just like in our project and [water-abstraction-ui](https://github.com/DEFRA/water-abstraction-ui) some of the internal legacy services also have an authorization strategy that applies [scope](https://hapi.dev/api/?v=21.3.3#-routeoptionsauthaccessscope) to their routes. We authenticate with them using the shared JWT token the previous team left behind. But that is not enough for those routes which have, for example, 'billing' configured as their authorization scope.

They expect the legacy UI to also let them know who the user is behind the request being made so they can confirm they have authorization to do that 'thing'. As this has already been checked by the UI it is pointless and wasteful. But it is what it is!

**water-abstraction-service** for example appears to have a couple of strategies; decoding the user email from the JWT token being one of them. But we suspect this was abandoned in favour of adding the user's `userId` as a header in the request.

If the header exists, the downstream service will extract the ID from it, make a call to [water-abstraction-tactical-idm](https://github.com/DEFRA/water-abstraction-tactical-idm) and then confirm the user and their scope. If they match the request is accepted.

All this means we need to update our `LegacyRequestLib` to support adding the `defra-internal-user-id` header to requests that need it.