Migrate Charging Module services to *.api.js
https://eaflood.atlassian.net/browse/WATER-4375

> Part of a series of changes related to replacing the create bill run journey to incorporate changes for two-part tariff

We know as part of the 2pT create bill run journey changes that we'll now be responsible for triggering any bill runs in the legacy service. This means we'll need to make a new call to the legacy service.

We already make one to each legacy app's `/health/info` endpoint and one to the `/refresh` bill run in[water-abstraction-service](https://github.com/DEFRA/water-abstraction-service). We were just going to copy and paste what we'd done before when we realised we had conflicting conventions.

Any calls to the [Charging Module API](https://github.com/DEFRA/sroc-charging-module.api) are wrapped in their own dedicated module. This removes duplication, simplifies testing and makes the intent much clearer.

All calls using the `LegacyRequestLib` are being done directly. This means we already have duplication in the billing engines and a conflicting pattern.

We want to make the Legacy stuff consistent with the Charging Module. Problem is, we're already seeing an explosion in services in the `app/services` as this is going to add another sub grouping. Plus, these services are different to the others as they are dedicated to making API requests to other services.

So, rather than moving Legacy to the Charging Module pattern, we're going to move the CHA to a new top level grouping called `app/api`. Files will be renamed from `*.service.js` to `*.api.js` and the call will be renamed from `go()` to `send()`.

Hopefully, that will make the intent of these types of modules clearer.