Move all effected files
Would have been nice to break this up but it broke so much stuff in the process it was clearer and easier just to do it en-masse and then confirm app was running and tests were passing again.

Also includes some cleanup of the JSDoc comments to bring them all to the same level of consistency.