Move legacy refresh requests to module
https://eaflood.atlassian.net/browse/WATER-4375

> Part of a series of changes related to replacing the create bill run journey to incorporate changes for two-part tariff

In [Migrate Charging Module services to *.request.js](https://github.com/DEFRA/water-abstraction-system/pull/797) we made structural changes to where the code that sends requests to 3rd party apps sit. This was in preparation for this change. For each request type we send to the [Charging Module API](https://github.com/DEFRA/sroc-charging-mdoule-api) we have a module. This removes duplication, centralises how that type of request is made and makes the intent clearer.

But legacy requests we're just using `Legacy.request.js` directly. We want a consistent pattern for how these things are done, especially as we are on the cusp of adding a new legacy request type. So, having moved the Charging Module requests to the pattern we want to adopt, it is now time to move how we are sending requests to the Legacy apps to refresh a bill run.